docs: remove the proxy from the sdks overview table
## What

This change removes the "unleash proxy" column from the SDK compat
table.

## Why

Because the proxy isn't technically an SDK. Having the column there
can be misleading to users.

In the latest case, a user pointed out that the table says the proxy
supports `isEnabled` with fallback values, but that there is no way to
use this via the API (or via client-side sdks) at the moment.